how it's possible to write for about 7 hours and have almost nothing done in the end.

+++ ServiceLayer/StockServices/AddNewGoodsService.cs
@@ -69,7 +69,7 @@ public class AddNewGoodsService(MusicalShopDbContext context, ISpecificTypeServi
             };
             goods.Description = dto.Description;
             goods.Name = dto.GoodsName;
-            goods.Price = dto.Price;
+            goods.Price = (int)dto.Price;
             goods.ReceiptDate = receiptDate;
             goods.SoftDeleted = false;
             goods.Status = dto.Status;
